% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{calcAlphaDiv}
\alias{calcAlphaDiv}
\title{Calculate \eqn{\alpha}-diversity Statistic}
\usage{
calcAlphaDiv(output, l = 0, covProfile = NULL)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}}

\item{l}{(numeric, default=0) Parameter l the diversity weight parameter.
If 0, the Shannon diversity index is calculated, if 1, the count diversity
index is calculated. Closer to 1 gives more weight to rare categories.}

\item{covProfile}{(numeric vector, default=NULL) Covariate profile. If NULL,
the baseline profile is used (ie all zeros).}
}
\value{
A matrix of \eqn{\alpha_p(t)} samples
}
\description{
Calculates the \eqn{\alpha_p(t)} statistic using the
Hill diversity index  from the \code{FunCZIDM} samples.
}
