% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{calcDeltaRA}
\alias{calcDeltaRA}
\title{Calculate the Multipicative Difference in Relateive Abundance}
\usage{
calcDeltaRA(output, change, forCovs, covProfile = NULL, forCats = NULL)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}}

\item{change}{(numeric vector) The change of the covariate.}

\item{forCovs}{(integer vector) Indices of \code{covariates}
for which \eqn{\Delta_vRA_{jp}(t)} to be calculated.}

\item{covProfile}{(numeric vector, default=NULL) Covariate profile.
If NULL, the baseline profile is used (ie all zeros).}

\item{forCats}{(integer vector, default=NULL) Indices of categories
for which \eqn{\Delta_vRA_{jp}(t)} is calculated. If NULL, then all
categories are used.}
}
\value{
A list of 3-dimensional array \eqn{\Delta_vRA_{jp}(t)} samples
}
\description{
This function calculates the \eqn{\Delta_vRA_{jp}(t)} statistic
from the samples of \code{FunCZIDM}.
}
