% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{calcRA}
\alias{calcRA}
\title{Calculates the Relateive Abundance}
\usage{
calcRA(output, covProfile = NULL)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}}

\item{covProfile}{(numeric vector, default=NULL) Covariate profile (default
is baseline, i.e. all zeros).}
}
\value{
A 3-dimensional array of \eqn{RA_{j}(t)} samples.
}
\description{
This function calculates the \eqn{RA_{j}(t)} from the from the
output of \code{FunCZIDM}
}
