% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{combineOutputs}
\alias{combineOutputs}
\title{Combine \code{FunCZIDM} Outputs}
\usage{
combineOutputs(
  outputFiles,
  saveToFile = TRUE,
  fileName = "combinedOutputs.rds"
)
}
\arguments{
\item{outputFiles}{(Character vector) Character vector of file paths to the
saved outputs.}

\item{saveToFile}{(logical, default=TRUE) Whether to save output to file.}

\item{fileName}{(character, default="combinedOutputs.rds") File name
for output.}
}
\value{
If \code{saveToFile} is TRUE, returns NULL after saving the combined
output to file. If \code{saveToFile} is FALSE, returns the combined output list.
}
\description{
Combines multiple FunCZIDM output files into a single output
list.
}
