% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM-package.R
\docType{package}
\name{FunCZIDM-package}
\alias{FunCZIDM-package}
\title{FunCZIDM: Functional Concurrent Zero-Inflated Dirichlet-Multinomial Regression}
\description{
Implements MCMC for the FunC-ZIDM model and helpers for simulation,
fitting, and visualization.
}
\details{
\subsection{Getting started}{
\itemize{
\item Install: see \code{README} (GitHub link below).
\item Main function: \code{?FunCZIDM}.
\item Vignette: see \code{vignette("FunCZIDMIntro")} (optional, install \code{ggtern}), or view online at https://brodyee.github.io/packageVignette/FunCZIDMIntro.html
\item Full index: \code{help(package = "FunCZIDM")}
}
}

\subsection{Links}{
\itemize{
\item GitHub: https://github.com/brodyee/FunCZIDM
}
}
}
\keyword{internal}
