% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{FunCZIDM}
\alias{FunCZIDM}
\title{FunCZIDM}
\usage{
FunCZIDM(
  counts,
  covariates,
  ids,
  varyingCov,
  iter = 10000,
  burnIn = 5000,
  thin = 1,
  adjustFreq = 250,
  returnBurnIn = FALSE,
  printProgress = TRUE,
  toReturn = NULL,
  betaInitial = NULL,
  rInitial = NULL,
  priors = NULL,
  proposalVars = NULL,
  covWithoutVC = NULL,
  df = 4,
  degree = 3,
  basisFunc = splines::bs,
  saveToFile = TRUE,
  fileName = "output.rds",
  saveNullInBetaCI = TRUE,
  nullInBetaCIFileName = "nullInBetaCI.csv"
)
}
\arguments{
\item{counts}{(integer matrix) Each categories counts for each sample.}

\item{covariates}{(data.frame) The covariates used for the regression,
should not include the intercept. These should correspond to the columns of
\code{counts}.}

\item{ids}{(integer vector) The index of subject ids in the 'covariates'.}

\item{varyingCov}{(numeric vector) The index of the covariate the varying
coefficients will be a function of in \code{covariates}. This is usually
time.}

\item{iter}{(integer, default=10000) Total number of iterations.}

\item{burnIn}{(integer, default=5000) Number of burn-in iterations, must be
less than \code{iter}.}

\item{thin}{(integer, default=1) Number of iterations to thin by.}

\item{adjustFreq}{(numeric, default=250) Frequency at which
\eqn{\beta_{jpd}^*} proposals are adjusted.}

\item{returnBurnIn}{(logical, default=FALSE) Whether to return burn-in
samples.}

\item{printProgress}{(logical, default=TRUE) Whether to print progress bar.}

\item{toReturn}{(vector, default=NULL) Vector of parameter names to
return in addition to the default parameters. The default parameters are:
\itemize{
\item{"beta"}: The \eqn{\beta_{jpd}^*} coefficients.
\item{"betaAcceptProp"}: The acceptance proportion for the
\eqn{\beta_{jpd}^*} proposals.
\item{"rAcceptProp"}: The acceptance proportion for the \eqn{r_{jp}}
proposals.
\item{"rMeans"}: The mean of the \eqn{r_{jp}} coefficients.
\item{"etaMeanPropZeros"}: The average level of zero-inflation indicated.
}
Returnable parameters are:
\itemize{
\item{"eta"}: The zero-inflation indicator.
\item{"c"}: The zero-inflation concentration parameter.
\item{"u"}: The auxiliary variable.
\item{"r"}: The individual- and category-specific coefficients.
\item{"phi"}: The individual- and category-specific coefficients sd
\item{"lambda"}: The local parameter of the horseshoe prior.
\item{"nu"}: The local parameter auxiliary variable of the horseshoe prior.
\item{"tau"}: The global parameter of the horseshoe prior.
\item{"xi"}: The global parameter auxiliary variable of the horseshoe prior.
\item{"RA"}: The mean estimated relative abundance of each sample.
}
If NULL, only the default parameters are returned.}

\item{betaInitial}{(matrix, default=NULL) Initialization values for
\eqn{beta_{jpd}^*}. If NULL, the initial values will be drawn from
Unif(-.75,.75).}

\item{rInitial}{(matrix, default=NULL) Initialization values for
\eqn{r_{jp}}. If NULL, the initial values will be drawn from Unif(-.05, .05).}

\item{priors}{(list, default=NULL) List with prior hyperparameters
specifications. The hyperparameters that can be specified are:
\itemize{
\item{"first beta sd"}: The prior sd for the first
\eqn{\beta_{jpd}^*} coefficient (\eqn{\beta_{j00}^*}).
\item{"a"}: The prior shape parameter for the variance of the \eqn{r_{jp}}.
\item{"b"}: The prior rate parameter for the variance of the \eqn{r_{jp}}.
\item{"alpha"}: The prior fisrt shape parameter for the probability of the
zero-inflation indicator.
\item{"beta"}: The prior second shape parameter for the probability of the
zero-inflation indicator.
\item{"kappaShape"}: The prior shape parameter for the \eqn{\kappa_{jp}},
i.e. the student-t portion of the regularized horseshoe.
\item{"kappaRate"}: The prior rate parameter for the \eqn{\kappa_{jp}},
i.e. the student-t portion of the regularized horseshoe.
}
NULL gives the default values, which are:
\itemize{
\item{"first beta sd"}: 1
\item{"a"}: 3
\item{"b"}: 9
\item{"alpha"}: .01
\item{"beta"}: 10
\item{"kappaShape"}: 100
\item{"kappaRate"}: 900
}}

\item{proposalVars}{(list, default=NULL) List with proposal
sd for \eqn{\beta_{jpd}^*} and \eqn{r_{jp}}. If NULL, the default proposal
sd are used, which are:
\itemize{
\item{"beta proposal sd"}: .1
\item{"r proposal sd"}: 1
}}

\item{covWithoutVC}{(vector, default=NULL) Indices of the \code{covariates}
that will not have varying coefficients. If NULL, all covariates will have
varying coefficients. Intercept will always have varying coefficients.}

\item{df}{(numeric, default=4) Degrees of freedom for the basis function.}

\item{degree}{(numeric, default=3) Degree for the basis function.}

\item{basisFunc}{(function, default=splines::bs) Function to generate basis
functions.}

\item{saveToFile}{(logical, default=TRUE) Whether to save output to file.}

\item{fileName}{(character, default="output.rds") File name for
output.}

\item{saveNullInBetaCI}{(logical, default=TRUE) Whether to a csv that
contains the proportion of the \eqn{\beta_{jp}(t)} coefficients that contain
the null value in their credible interval. Gives an idea of which covariate
and category combinations may have meaningful effects.}

\item{nullInBetaCIFileName}{(character, default="nullInBetaCI.csv")
File name for null CI statistics.}
}
\value{
Depending on saveToFile, either returns NULL after saving or an
output list. The output list contains:
\itemize{
\item{"beta"}: The \eqn{\beta_{jpd}^*} samples.
\item{"betaAcceptProp"}: The acceptance proportion for the
\eqn{\beta_{jpd}^*} proposals after burn-in.
\item{"rAcceptProp"}: The acceptance proportion for the \eqn{r_{jp}}
proposals after burn-in.
\item{"rMeans"}: The mean of the \eqn{r_{jp}} samples.
\item{"etaMeanPropZeros"}: The average level of zero-inflation indicated.
\item{"interiorKnots"}: The interior knots of the basis functions.
\item{"boundaryKnots"}: The boundary knots of the basis functions.
\item{"df"}: The degrees of freedom used for the basis functions.
\item{"basisFunc"}: The basis function used.
\item{"varyingCov"}: The values of the covariate that the varying
coefficients are a function of.
\item{"colMapping"}: A list mapping the covariate indices to the column names
of the design matrix.
\item{"XvartoXColMapping"}: A vector mapping the columns of the design matrix
with the basis design matrix.
\item{"catNames"}: The category names from the counts.
\item{"centerScaleList"}: A list containing the center and scale of each
continuous covariate. If a covariate is a factor, then "no c/s" is stored.
}
}
\description{
This function is used to sample from the FunC-ZIDM regression
model.
}
\examples{
\dontrun{
# toy example with included data
library(FunCZIDM)

data(infantData)
# Convert the categorical variables to factors
infantCovariates$gender <- as.factor(infantCovariates$gender)
infantCovariates$mode.of.birth <- as.factor(infantCovariates$mode.of.birth)
infantCovariates$Room.category <- as.factor(infantCovariates$Room.category)
infantCovariates$milk <- as.factor(infantCovariates$milk)
infantCovariates$milk <- relevel(infantCovariates$milk, ref="< 10\%")
infantCovariates$Period.of.study <- as.factor(infantCovariates$Period.of.study)
infantCovariates$Period.of.study <- relevel(infantCovariates$Period.of.study,
                                            ref="before")

# getting the index of the subject id's and the time-varying variable
idIdx <- which(colnames(infantCovariates) == "Subject")
timeIdx <- which(colnames(infantCovariates) == "Day.of.life.sample.obtained")
# extracting the subject ids and tv variable
ids <- infantCovariates[, idIdx]
time <- infantCovariates[, timeIdx]
infantCovariates <- infantCovariates[, -c(idIdx, timeIdx), drop = FALSE]

idIdx <- which(colnames(infantCounts) == "Subject")
timeIdx <- which(colnames(infantCounts) == "Day.of.life.sample.obtained")
infantCounts <- infantCounts[, -c(idIdx, timeIdx), drop = FALSE]
infantCounts <- as.matrix(infantCounts)

output <- FunCZIDM(counts=infantCounts, 
                   covariates=infantCovariates, 
                   ids=ids,
                   varyingCov=time,
                   iter=85,
                   burnIn=75,
                   thin=10,
                   covWithoutVC=c(which(grepl("Period", colnames(infantCovariates)))),
                   saveToFile=FALSE,
                   saveNullInBetaCI=FALSE)
}
}
\seealso{
\code{\link{calcRA}},
\code{\link{calcDeltaRA}},
\code{\link{calcAlphaDiv}},
\code{\link{calcDeltaAlphaDiv}},
\code{\link[=FunCZIDM-package]{FunCZIDM-package}}
}
