% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{getBetaFunctions}
\alias{getBetaFunctions}
\title{Get \eqn{\beta_{jp}(t)} Functions}
\usage{
getBetaFunctions(output, cov = NULL)
}
\arguments{
\item{output}{(list) Output list returned from \code{FunCZIDM}}

\item{cov}{(integer, default=NULL) The index of \code{covariates} from
\code{FunCZIDM} for which \eqn{\beta_{jp}(t)} functions to extract.
Intercept if NULL.}
}
\value{
A 3-dimensional array of \eqn{\beta_{jp}(t)} function samples for
the specified covariate.
}
\description{
Extracts the \eqn{\beta_{jp}(t)} function samples from the
output of \code{FunCZIDM}
}
