% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{getMinMaxDeltaAlphaDiv}
\alias{getMinMaxDeltaAlphaDiv}
\title{Get Min/Max Multiplicative Change in Alpha Diversity}
\usage{
getMinMaxDeltaAlphaDiv(
  output,
  changeFrom,
  changeTo,
  forCov,
  covariates,
  varyingCov,
  varyingCovBreaks,
  l = 0,
  reqSig = FALSE
)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}.}

\item{changeFrom}{(numeric) Value to change from for the covariate.}

\item{changeTo}{(numeric) Value to change to for the covariate.}

\item{forCov}{(integer) Index of the covariate for which to calculate the change.}

\item{covariates}{(data.frame) Covariate data.}

\item{varyingCov}{(numeric vector) Values of the varying covariate.}

\item{varyingCovBreaks}{(numeric vector) Breaks for intervals of the varying covariate.}

\item{l}{(numeric, default=0) Diversity parameter for the Hill diversity index.}

\item{reqSig}{(logical, default=FALSE) If TRUE, only consider intervals where the 95\% CI does not include 1.}
}
\value{
A list with elements \code{maxCov} and \code{minCov}, each a matrix of covariate profiles corresponding to the maximum and minimum mean multiplicative change in alpha diversity for each interval.
}
\description{
Finds the covariate profiles corresponding to the minimum and
maximum mean multiplicative change in alpha diversity
(\eqn{\Delta_v\alpha_p(t)}) for a given covariate, within intervals of the
varying covariate.
}
