% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{getMinMaxDeltaRA}
\alias{getMinMaxDeltaRA}
\title{Get Min/Max Multiplicative Change in Relative Abundance}
\usage{
getMinMaxDeltaRA(
  output,
  changeFrom,
  changeTo,
  forCov,
  forCat,
  covariates,
  varyingCov,
  varyingCovBreaks,
  reqSig = FALSE
)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}.}

\item{changeFrom}{(numeric) Value to change from.}

\item{changeTo}{(numeric) Value to change to.}

\item{forCov}{(integer) Index of the covariate for which to calculate.}

\item{forCat}{(integer) Index of the category for which to calculate.}

\item{covariates}{(data.frame) Covariate data.}

\item{varyingCov}{(numeric vector) Values of the varying covariate.}

\item{varyingCovBreaks}{(numeric vector) Breaks for intervals of the varying covariate.}

\item{reqSig}{(logical, default=FALSE) If TRUE, only consider intervals where the 95\% CI does not include 1.}
}
\value{
A list with elements "maxCov" and "minCov", each a matrix of covariate profiles.
}
\description{
Finds the covariate profiles corresponding to the minimum and maximum mean multiplicative change in relative abundance (\eqn{\Delta_vRA_{jp}(t)}) for a given covariate and category, within intervals of the varying covariate.
}
