% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{getPropNullInCI}
\alias{getPropNullInCI}
\title{Get Proportion of 0 in 95\% CI of \eqn{\beta_{jp}(t)}}
\usage{
getPropNullInCI(output, varCovRange = NULL)
}
\arguments{
\item{output}{(list) Output list from \code{FunCZIDM}.}

\item{varCovRange}{(numeric vector, default=NULL) Range of the varying
covariate to evaluate the \eqn{\beta_{jp}(t)} functions. By default, the
range is the min and max of the varying covariate in the data.}
}
\value{
A matrix with columns "category", "covariate", and "Proportion 0 in
95 CI"
}
\description{
Calculates the proportion of \eqn{\beta_{jp}(t)} functions that
have the zero function in their 95\% credible interval.
}
