% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{runPlotApp}
\alias{runPlotApp}
\title{Runs the Shiny App for Visualization}
\usage{
runPlotApp(useCondaEnv = FALSE, ...)
}
\arguments{
\item{useCondaEnv}{(logical, default=FALSE) Whether to use the conda
environment for running the app. If TRUE, the conda environment named
"shinyPy" will be used. If FALSE, the python from reticulate will be used.}

\item{...}{Additional arguments to pass to \code{shinyAppDir}.}
}
\description{
Runs the Shiny app for visualizing the output of
\code{FunCZIDM}.
}
