% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM-package.R
\docType{package}
\name{FunCZIDM-package}
\alias{FunCZIDM}
\alias{FunCZIDM-package}
\title{FunCZIDM: Functional Concurrent Zero-Inflated Dirichlet Multinomial Regression}
\description{
This package contains functions to sample from FunC-ZIDM regression model, and has the nessicary functions to process the samples.
}
\author{
\strong{Maintainer}: Brody Erlandson \email{erlandsonbrody@gmail.com}

}
\keyword{internal}
