% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{FunCZIDM}
\alias{FunCZIDM}
\title{FunCZIDM}
\usage{
FunCZIDM(
  counts,
  covariates,
  ids,
  varyingCov,
  rCols = NULL,
  iter = 10000,
  burnIn = 5000,
  thin = 1,
  adjustFreq = 250,
  proposalCap = 0,
  ZIGrouped = TRUE,
  returnBurnIn = FALSE,
  printProgress = TRUE,
  toReturn = NULL,
  betaInitial = NULL,
  rInitial = NULL,
  priors = NULL,
  proposalVars = NULL,
  covWithVC = NULL,
  df = 4,
  degree = 3,
  basisFunc = splines::bs,
  saveToFile = TRUE,
  fileName = "output.rds",
  saveNullInBetaCI = TRUE,
  nullInBetaCIFileName = "nullInBetaCI.csv"
)
}
\arguments{
\item{counts}{(integer matrix) Each categories counts for each sample.}

\item{covariates}{(data.frame) The covariates used for the regression,
should not include the intercept. These should correspond to the columns of
\code{counts}.}

\item{ids}{(integer vector) The index of subject ids in the 'covariates'.}

\item{varyingCov}{(numeric vector) The index of the covariate the varying
coefficients will be a function of in \code{covariates}. This is usually time.}

\item{rCols}{(integer vector, default=NULL) Indices of \code{covariates} with
subject- and category- specific effects. By default, it will only be the
intercept. For this model, the intercept always has subject- and category-
specific effects, and does not need to be specified. Check the \code{VarCoZIDM}
function for a model without any subject- and category- specific effects.}

\item{iter}{(integer, default=10000) Total number of iterations.}

\item{burnIn}{(integer, default=5000) Number of burn-in iterations, must be
less than \code{iter}.}

\item{thin}{(integer, default=1) Number of iterations to thin by.}

\item{adjustFreq}{(numeric, default=250) Frequency at which
\eqn{\beta_{jpd}^*} proposals are adjusted.}

\item{proposalCap}{(numeric, default=0) Cap for proposal variances, if 0
there will be no cap.}

\item{ZIGrouped}{(logical, default=TRUE) Whether to use the zero inflation
indicator for all samples of an id (TRUE), or each sample have their own
indicator (FALSE).}

\item{returnBurnIn}{(logical, default=FALSE) Whether to return burn-in
samples.}

\item{printProgress}{(logical, default=TRUE) Whether to print progress bar.}

\item{toReturn}{(vector, default=NULL) Vector of parameter names to
return in addition to the default parameters. The default parameters are:
\itemize{
\item{"beta"}: The \eqn{\beta_{jpd}^*} coefficients.
\item{"betaAcceptProp"}: The acceptance proportion for the
\eqn{\beta_{jpd}^*} proposals.
\item{"rAcceptProp"}: The acceptance proportion for the \eqn{r_{jp}}
proposals.
\item{"rMeans"}: The mean of the \eqn{r_{jp}} coefficients.
\item{"etaMeanPropZeros"}: The average level of zero-inflation indicated.
}
Returnable parameters are:
\itemize{
\item{"eta"}: The zero-inflation indicator.
\item{"c"}: The zero-inflation concentration parameter.
\item{"u"}: The auxiliary variable.
\item{"r"}: The individual- and category-specific coefficients.
\item{"phi"}: The individual- and category-specific coefficients sd
\item{"lambda"}: The local parameter of the horseshoe prior.
\item{"nu"}: The local parameter auxiliary variable of the horseshoe prior.
\item{"tau"}: The global parameter of the horseshoe prior.
\item{"xi"}: The global parameter auxiliary variable of the horseshoe prior.
\item{"RA"}: The mean estimated relative abundance of each sample.
}
If NULL, only the default parameters are returned.}

\item{betaInitial}{(matrix, default=NULL) Initialization values for
\eqn{beta_{jpd}^*}. If NULL, the initial values will be drawn from
Unif(-.75,.75).}

\item{rInitial}{(matrix, default=NULL) Initialization values for
\eqn{r_{jp}}. If NULL, the initial values will be drawn from Unif(-.05, .05).}

\item{priors}{(list, default=NULL) List with prior hyperparameters
specifications. The hyperparameters that can be specified are:
\itemize{
\item{"first beta sd"}: The prior sd for the first
\eqn{\beta_{jpd}^*} coefficient (\eqn{\beta_{j00}^*}).
\item{"a"}: The prior shape parameter for the variance of the \eqn{r_{jp}}.
\item{"b"}: The prior scale parameter for the variance of the \eqn{r_{jp}}.
\item{"alpha"}: The prior fisrt shape parameter for the probability of the
zero-inflation indicator.
\item{"beta"}: The prior second shape parameter for the probability of the
zero-inflation indicator.
}
NULL gives the default values, which are:
\itemize{
\item{"first beta sd"}: 1
\item{"a"}: 3
\item{"b"}: 9
\item{"alpha"}: .01
\item{"beta"}: 10
}}

\item{proposalVars}{(list, default=NULL) List with proposal
sd for \eqn{\beta_{jpd}^*} and \eqn{r_{jp}}. If NULL, the default proposal
sd are used, which are:
\itemize{
\item{"beta proposal sd"}: .3
\item{"r proposal sd"}: 1
}}

\item{covWithVC}{(vector, default=NULL) Indices of the \code{covariates}
that will have varying coefficients. If NULL, all covariates will have
varying coefficients. Intercept will always have varying coefficients.}

\item{df}{(numeric, default=4) Degrees of freedom for the basis function.}

\item{degree}{(numeric, default=3) Degree for the basis function.}

\item{basisFunc}{(function, default=splines::bs) Function to generate basis
functions.}

\item{saveToFile}{(logical, default=TRUE) Whether to save output to file.}

\item{fileName}{(character, default="output.rds") File name for
output.}

\item{saveNullInBetaCI}{(logical, default=TRUE) Whether to a csv that
contains the proportion of the \eqn{\beta_{jp}(t)} coefficients that contain
the null value in their credible interval. Gives an idea of which covariate
and category combinations may have meaningful effects.}

\item{nullInBetaCIFileName}{(character, default="nullInBetaCI.csv")
File name for null CI statistics.}
}
\value{
Depending on saveToFile, either returns NULL after saving or an
output list.
}
\description{
This function is used to sample from the FunC-ZIDM regression
model.
}
