% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunCZIDM.R
\name{generateData}
\alias{generateData}
\title{Generate Data}
\usage{
generateData(n, c, p, totalCountRange = c(100, 200), numActive = 4)
}
\arguments{
\item{n}{(integer) Number of individuals.}

\item{c}{(integer) Number of categories.}

\item{p}{(integer) Number of functional covariates.}

\item{totalCountRange}{(numeric vector, default=c(100, 200)) Range of total
counts.}

\item{numActive}{(integer, default=4) Number of categories with active
covariates. Alway 1:numActive categories are active.}
}
\value{
A list containing the generated data: counts, covariates, ids,
timepoints and the true parameters.
}
\description{
Generates synthetic data for testing the model.
}
